refactor AuthorizationClient into new file
